backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic[email]>=2.5.0
python-multipart>=0.0.6
opencv-python>=4.8.1
numpy>=1.26.2
torch
torchvision
ffmpeg-python>=0.2.0
pillow>=10.1.0
boto3>=1.29.7
celery>=5.3.4
redis>=5.0.1
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
PyJWT>=2.8.0
email-validator>=2.0.0
python-dotenv>=1.0.0